Final

See merge request sandip125/ward-robe!21

+++ src/main/java/nce/majorproject/services/impl/CommentServiceImpl.java
@@ -165,7 +165,7 @@ public class CommentServiceImpl implements CommentService {
             log.info("No rating found");
         }
         if(reviewRating!=null){
-        return  reviewRating.getRating();
+        return  (float)reviewRating.getRating();
         }
         return 0;
     }
